When Swagger JSON contains header definition, it will be placed in the kwargs JSON object.
Because of the enhancement that was made to allow specification of the
API Key thru environmental variable as header, the side-effect is there
are two headers being sent to the requests.request() call.

To resolve the issue, we can check if header exist in kwargs JSON
object, if so then we merge both headers into a single combined header
then pass it as an input to requests.request() call.